Only parse one feed with JsonVideoParser
JsonVideoParser takes a sequence of futures each containing a string of
the json feed and returns a sequence of future lists of videos. It would
be better to have the JsonVideoParser only parse one feed and do it
directly, a future can map over the JsonVideoParser if necessary. This
commit changes that.